Fix for ch sample audio format

+++ applications/ffd/inference_engines/api/inference_engine.h
@@ -8,6 +8,6 @@
 
 /* Generic interface for inference engines */
 int32_t inference_engine_create(uint32_t priority, void *args);
-int32_t inference_engine_sample_push(uint8_t *buf, size_t bytes);
+int32_t inference_engine_sample_push(int32_t *buf, size_t bytes);
 
 #endif /* INFERENCE_ENGINE_H_ */
